requirements.txt
python-telegram-bot
python-constraint
openai
